backend/requirements.txt
affine==2.4.0
aiofiles==24.1.0
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
boto3==1.34.142
botocore==1.34.142
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
dnspython==2.6.1
docker==7.1.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
fonttools==4.53.1
frozenlist==1.4.1
GDAL==3.6.2
GeoAlchemy2==0.15.1
google-pasta==0.2.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib-metadata==6.11.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
# numpy==1.23.5  # Compatible version with rasterio\
# numpy==1.24.4
opencv-python==4.10.0.84
orjson==3.10.4
packaging==24.1
pandas==2.2.2
pathos==0.3.2
pillow==10.4.0
platformdirs==4.2.2
pox==0.3.4
ppft==1.7.6.8
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rasterio==1.3.8
referencing==0.35.1
replicate==0.30.1
reportlab==4.2.2
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
s3transfer==0.10.2
sagemaker==2.226.1
schema==0.7.7
setuptools==70.3.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
snuggs==1.4.7
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
starlette==0.37.2
tblib==3.0.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.15  # Compatible version with botocore
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
